feat: 大规模 ERP 系统扩展 - 财务模块、前端多端应用迭代、E2E 测试
后端新增模块：
- finance: AccountSubject/BusinessAccounting/FinancialReport/Ledger/Payable/Receivable/Voucher 完整财务模块
- price-engine: PriceTier 分层定价计算器
- purchase/sales: 集成配置与服务

前端新增模块：
- pc-admin: budget/finance/fixed-asset/pricing-tiers/tenant 等业务视图
- pc-admin: GlobalSearch/TabsView/FileAttachment/ProductSelector 等通用组件
- pc-admin: useDraftAutoSave/shortcut/columnStorage 等工具
- mobile-admin: about/expense/leave/more/profile/scan 等移动端视图
- driver-delivery: components 类型定义

前端重构：
- pc-admin: CRM/ERP/Sales/Purchase/Stock 等核心业务页面大规模重构
- pc-admin: BasicLayout/router/guard/request/logger 等基础设施优化
- packages/components: ARForm/ARFormItem/ARTable 表单表格组件增强

E2E 测试：
- approval-flow/purchase-flow/stock-flow 业务流程测试

其他：
- docs/gap-analysis.md: 差距分析文档
- pnpm-lock.yaml: 依赖更新

+++ backend/erp/finance/src/main/java/cn/aiedge/erp/finance/controller/FinanceAccountController.java
@@ -11,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -70,7 +71,7 @@ public class FinanceAccountController {
     @PutMapping("/balance/{accountId}")
     public Result<Void> updateAccountBalance(
             @Parameter(description = "账户ID") @PathVariable Long accountId,
-            @Parameter(description = "金额") @RequestBody Double amount) {
+            @Parameter(description = "金额") @RequestBody BigDecimal amount) {
         boolean success = financeAccountService.updateAccountBalance(accountId, amount);
         return success ? Result.success() : Result.error("更新失败");
     }
